Trainings and tournaments

+++ IzvorniKod/API/models/TournamentModel.js
@@ -20,7 +20,7 @@ module.exports = class Tournament {
             for(let i = 0; i < results.length; i++){
                 let temp = new Tournament(results[i].trainer_id, results[i].tournamentstarttimedate, results[i].tournamentdurationmin);
                 temp.id = results[i].id;
-                temp.participantsNo = results[i].participantsNo;
+                temp.participantsNo = results[i].participantsno;
                 temp.showing = results[i].showing;
                 allTournaments[i] = temp;
             }
@@ -35,11 +35,7 @@ module.exports = class Tournament {
         let result = [];
         if(tournaments?.length != 0){
             for(let i = 0; i < tournaments.length; i++){
-                let newTournament = new Tournament(tournaments[i].trainer_id, tournaments[i].tournamentstarttimedate, tournaments[i].tournamentdurationmin);
-                newTournament.id = tournaments[i].id;
-                newTournament.participantsNo = tournaments[i].participantsNo;
-                newTournament.showing = tournaments[i].showing;
-                result[i] = newTournament;
+                result[i] = tournaments[i];
             }
             console.log(result);
             return result;
@@ -53,7 +49,7 @@ module.exports = class Tournament {
         if(results?.length == 1){
             let tournament = new Tournament(results[0].trainer_id, results[0].tournamentstarttimedate, results[0].tournamentdurationmin);
             tournament.id = results[0].id;
-            tournament.participantsNo = results[0].participantsNo;
+            tournament.participantsNo = results[0].participantsno;
             tournament.showing = results[0].showing;
             return tournament;
         }
@@ -69,17 +65,15 @@ module.exports = class Tournament {
     }
 
     async signupForTournament(userId) {
-        console.log(this.participantsNo);
         this.participantsNo++;
-        console.log(this.participantsNo);
         await dbUpdateParticipantsNo(this);
         await dbSignupForTournament(this, userId);
     }
 
-    async cancelTournament() {
+    async cancelTournament(userId) {
         this.participantsNo--;
         await dbUpdateParticipantsNo(this);
-        await dbCancelTournament(this);
+        await dbCancelTournament(this, userId);
     }
 
     async persist() {
@@ -108,7 +102,7 @@ const dbGetAllTournaments = async() => {
 
 
 const dbGetAllScheduledTournaments = async (id)=>{
-    const sql = "select * from scheduledTournament join users on member_id = id where member_id =" + id;
+    const sql = "select * from scheduledTournament  where member_id = '" + id + "'";
 
     try {
         const result = await db.query(sql, []);
@@ -196,8 +190,8 @@ const dbUpdateParticipantsNo = async (tournament) => {
 }
 
 
-const dbCancelTournament = async (tournament) => {
-    const sql = "delete from scheduledtournament where tournament_id = '" + tournament.id + "'";
+const dbCancelTournament = async (tournament, userId) => {
+    const sql = "delete from scheduledtournament where tournament_id = '" + tournament.id + "' AND member_id = '" + userId+"'";
 
     try {
         await db.query(sql, []);
